requirements.txt
autopep8==2.0.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
docopt==0.6.2
experta==1.9.4
Flask==2.2.2
Flask-Cors==3.0.10
frozendict==1.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
mysql-connector-python==8.0.31
numpy==1.23.4
pandas==1.5.1
pipreqs==0.4.11
protobuf==3.20.1
pycodestyle==2.9.1
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
schema==0.6.7
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
threadpoolctl==3.1.0
tomli==2.0.1
urllib3==1.26.12
Werkzeug==2.2.2
yarg==0.1.9
zipp==3.10.0
